requirements.txt
gradio==3.32.0
transformers==4.34.0
torch==2.0.1
opencv-python==4.7.0.72
pillow==9.5.0
numpy==1.24.4
requests==2.31.0
